+++ ovp_projects/views/project.py
@@ -1,14 +1,10 @@
-from django.core.exceptions import ObjectDoesNotExist
 from django.db.models import Q
 
 from ovp_projects.serializers import project as serializers
-from ovp_projects.serializers.apply import ApplyCreateSerializer, ApplyRetrieveSerializer
 from ovp_projects import models
 from ovp_projects import helpers
 from ovp_projects.permissions import ProjectCreateOwnsOrIsOrganizationMember
 from ovp_projects.permissions import ProjectRetrieveOwnsOrIsOrganizationMember
-
-from ovp_users import models as users_models
 
 from rest_framework import decorators
 from rest_framework import mixins
@@ -26,6 +22,9 @@ class ProjectResourceViewSet(mixins.CreateModelMixin, mixins.RetrieveModelMixin,
   lookup_field = 'slug'
   lookup_value_regex = '[^/]+' # default is [^/.]+ - here we're allowing dots in the url slug field
 
+  ##################
+  # ViewSet routes #
+  ##################
   def create(self, request, *args, **kwargs):
     request.data['owner'] = request.user.pk
 
@@ -35,7 +34,6 @@ class ProjectResourceViewSet(mixins.CreateModelMixin, mixins.RetrieveModelMixin,
 
     headers = self.get_success_headers(serializer.data)
     return response.Response(serializer.data, status=status.HTTP_201_CREATED, headers=headers)
-
 
   def partial_update(self, request, *args, **kwargs):
     """ We do not include the mixin as we want only PATCH and no PUT """
@@ -65,58 +63,10 @@ class ProjectResourceViewSet(mixins.CreateModelMixin, mixins.RetrieveModelMixin,
     serializer = self.get_serializer_class()(projects, many=True, context=self.get_serializer_context())
     return response.Response(serializer.data)
 
-  @decorators.detail_route(['POST'])
-  def apply(self, request, *args, **kwargs):
-    data = request.data
-    data.pop('user', None)
 
-    project = self.get_object()
-    data['project'] = project.id
-
-    if request.user.is_authenticated():
-      user = request.user
-      data['username'] = user.name
-      data['email'] = user.email
-      data['phone'] = user.phone
-      data['user'] = user.id
-
-    try:
-      existing_apply = models.Apply.objects.get(project=project, email=data['email'], canceled=True)
-      existing_apply.canceled = False
-      existing_apply.save()
-    except ObjectDoesNotExist:
-      apply_sr = self.get_serializer_class()(data=data, context=self.get_serializer_context())
-      apply_sr.is_valid(raise_exception=True)
-      apply_sr.save()
-
-    return response.Response({'detail': 'Successfully applied.'}, status=status.HTTP_200_OK)
-
-  @decorators.detail_route(['POST'])
-  def unapply(self, request, *args, **kwargs):
-    project = self.get_object()
-    user = request.user
-
-    try:
-      existing_apply = models.Apply.objects.get(project=project, email=user.email, canceled=False)
-      existing_apply.canceled = True
-      existing_apply.save()
-    except ObjectDoesNotExist:
-      return response.Response({'detail': 'This is user is not applied to this project.'}, status=status.HTTP_400_BAD_REQUEST)
-
-    return response.Response({'detail': 'Successfully unapplied.'}, status=status.HTTP_200_OK)
-
-  @decorators.detail_route(['GET'])
-  def applies(self, request, *arg, **kwargs):
-    project = self.get_object()
-    applies = models.Apply.objects.filter(project=project)
-
-    serializer = self.get_serializer_class()(applies, many=True, context=self.get_serializer_context())
-
-    return response.Response(serializer.data)
-
-
-  # We need to override get_permissions and get_serializer_class to work
-  # with multiple serializers and permissions
+  ###################
+  # ViewSet methods #
+  ###################
   def get_permissions(self):
     request = self.get_serializer_context()['request']
     if self.action == 'create':
@@ -125,17 +75,8 @@ class ProjectResourceViewSet(mixins.CreateModelMixin, mixins.RetrieveModelMixin,
       else:
         self.permission_classes = (permissions.IsAuthenticated, ProjectCreateOwnsOrIsOrganizationMember)
 
-    if self.action in ['applies', 'partial_update']:
+    if self.action == 'partial_update':
       self.permission_classes = (permissions.IsAuthenticated, ProjectRetrieveOwnsOrIsOrganizationMember)
-
-    if self.action == 'unapply':
-      self.permission_classes = (permissions.IsAuthenticated, )
-
-    if self.action == 'apply':
-      if helpers.get_settings().get('UNAUTHENTICATED_APPLY', False):
-        self.permission_classes = ()
-      else:
-        self.permission_classes = (permissions.IsAuthenticated, )
 
     if self.action == 'retrieve':
       self.permission_classes = ()
@@ -151,10 +92,6 @@ class ProjectResourceViewSet(mixins.CreateModelMixin, mixins.RetrieveModelMixin,
   def get_serializer_class(self):
     if self.action in ['create', 'partial_update']:
       return serializers.ProjectCreateUpdateSerializer
-    if self.action in ['apply', 'unapply']:
-      return ApplyCreateSerializer
-    if self.action == 'applies':
-      return ApplyRetrieveSerializer
     if self.action == 'manageable':
       return serializers.ProjectRetrieveSerializer
     if self.action == 'close':