+++ CHANGES.rst
@@ -175,3 +175,7 @@ v1.1.11[unreleased]
 -----------
 * Apply.status is now a delimited choice field
 * Apply status returned by API is not an "key" field anymore, but a readable string, so instead of getting "applied" and "unapplied" you might now get "Applied" and "Canceled".
+* Apply and unapply routes are changed:
+/projects/{slug}/apply => /projects/{slug}/applies/apply
+/projects/{slug}/unapply => /projects/{slug}/applies/unapply
+